Request and export all PRs of a milestone to a JSON file
Providing the repository owner and name as well as the name of the milestone of interest, this script sends a request using the GitHub API to retrieve all the closed pull requests associated to a milestone.

If the request is successful, the response is exported as a JSON file, without any parsing phase.

If the response spans over several pages, all the pages are retrieved and exported to the final JSON file, with respect to the sorting order (if unspecified: most recently updated first).