Add a release note formatter
Given the response from the GitHub API provided as a JSON file, the formatter parses all the pull requests associated to a milestone and outputs a release note in markdown.

All pull requests that were cancelled (closed but not merged) are ignored. The list of contributors can be generated in a separate file.

Labels to exclude from the final release note can be provided and they will be written in separate files. Several labels can be excluded at once, and they can also be concatenated. Concatened labels will be dumped in the same file, while stand-alone labels will each be dumped into their own file.