Add support for the --label-include option
Similarly to the "--label-exclude" option, "--label-include" identifies merged pull requests that have been tagged with at least one of the provided labels.

Instead of removing them fully from the final release note, they are removed from the main list and placed into subsections instead.

A label can be given both as an "include" and "exclude" label at the same time: it will be part of the release note in its own subsection, and a separate markdown file containing all the pull requests that were tagged with will be dumped.