Add options to include or exclude words
Similarly to --label-include and --label-exclude, two new options, --word-include and --word-exclude, are added to filter pull requests based on the presence of at least one of the provided words in the their title.

Pull requests that contain these words will be included in the release note or excluded from it in the very same fashion as they are with labels.

N.B: these options were added with a lot of code duplication. Refactoring will be necessary.