Add baseline to find more info about authors
Add a baseline scripts that, given a list of pull requests or issues in a JSON file, requests user information once for every unique user that authored a pull request or an issue.

The goal will be to extract the full name associated to the username, and maybe some additional information.